fix: Clone config in editor. Fixes #233
This _should_ fix the issue with frozen objects.
After weighing the options I landed on simply JSON parse/stringify as
the config only holds JSON values and should never be able to hold
cyclic values.
If this turns out to not be correct in the future the logic should be
replaced with a library that does proper deep clone